requirements.txt
# Core OpenAI
openai>=1.0.0

# LangChain packages (updated for new module structure)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1

# Vector Database
chromadb>=0.4.0

# Document Processing
pypdf>=3.0.0

# Web UI
streamlit>=1.28.0

# Configuration
python-dotenv>=1.0.0

# PDF Generation (optional)
reportlab>=4.0.0
fpdf2>=2.7.0
